move CSCObjectMap, CSCStationIndex, DTObjectMap, DTStationIndex to their own files
in the process:
  - rename ObjectMap to DTObjectMap
  - rename ObjectMapCSC to CSCObjectMap
  - merge multiple copies
