+++ RecoLocalMuon/RPCRecHit/interface/DTSegtoRPC.h
@@ -25,47 +25,4 @@ private:
   std::vector<uint32_t> extrapolatedRolls;
 };
 
-class DTStationIndex{
-public: 
-  DTStationIndex():_region(0),_wheel(0),_sector(0),_station(0){}
-  DTStationIndex(int region, int wheel, int sector, int station) : 
-    _region(region),
-    _wheel(wheel),
-    _sector(sector),
-    _station(station){}
-  ~DTStationIndex(){}
-  int region() const {return _region;}
-  int wheel() const {return _wheel;}
-  int sector() const {return _sector;}
-  int station() const {return _station;}
-  bool operator<(const DTStationIndex& dtind) const{
-    if(dtind.region()!=this->region())
-      return dtind.region()<this->region();
-    else if(dtind.wheel()!=this->wheel())
-      return dtind.wheel()<this->wheel();
-    else if(dtind.sector()!=this->sector())
-      return dtind.sector()<this->sector();
-    else if(dtind.station()!=this->station())
-      return dtind.station()<this->station();
-    return false;
-  }
-
-private:
-  int _region;
-  int _wheel;
-  int _sector;
-  int _station; 
-};
-
-class ObjectMap{
-public:
-  static ObjectMap* GetInstance(const edm::EventSetup& iSetup);
-  std::set<RPCDetId> GetRolls(DTStationIndex dtstationindex){return mapInstance->rollstoreDT[dtstationindex];}
-//protected:
-  std::map<DTStationIndex,std::set<RPCDetId> > rollstoreDT;
-  ObjectMap(const edm::EventSetup& iSetup);
-private:
-  static ObjectMap* mapInstance;
-}; 
-
 #endif
