+++ RecoLocalMuon/RPCRecHit/interface/TracktoRPC.h
@@ -51,8 +51,7 @@
 #include "TrackingTools/TrackRefitter/interface/TrackTransformerForCosmicMuons.h"
 #include "TrackingTools/TrackRefitter/interface/TrackTransformerBase.h"
 #include "TrackingTools/TrackRefitter/interface/TrackTransformer.h"
-//#include "RecoLocalMuon/RPCRecHit/interface/DTSegtoRPC.h"
-//#include "RecoLocalMuon/RPCRecHit/interface/CSCSegtoRPC.h"
+
 
 using reco::MuonCollection;
 using reco::TrackCollection;
@@ -76,90 +75,5 @@ private:
  TrackTransformerBase *theTrackTransformer;
  edm::ESHandle<Propagator> thePropagator;
 };
-
-class DTStationIndex2{
-public: 
-  DTStationIndex2():_region(0),_wheel(0),_sector(0),_station(0){}
-  DTStationIndex2(int region, int wheel, int sector, int station) : 
-    _region(region),
-    _wheel(wheel),
-    _sector(sector),
-    _station(station){}
-  ~DTStationIndex2(){}
-  int region() const {return _region;}
-  int wheel() const {return _wheel;}
-  int sector() const {return _sector;}
-  int station() const {return _station;}
-  bool operator<(const DTStationIndex2& dtind) const{
-    if(dtind.region()!=this->region())
-      return dtind.region()<this->region();
-    else if(dtind.wheel()!=this->wheel())
-      return dtind.wheel()<this->wheel();
-    else if(dtind.sector()!=this->sector())
-      return dtind.sector()<this->sector();
-    else if(dtind.station()!=this->station())
-      return dtind.station()<this->station();
-    return false;
-  }
-
-private:
-  int _region;
-  int _wheel;
-  int _sector;
-  int _station; 
-};
-
-class ObjectMap2{
-public:
-  static ObjectMap2* GetInstance(const edm::EventSetup& iSetup);
-  std::set<RPCDetId> GetRolls(DTStationIndex2 dtstationindex){return mapInstance->rollstoreDT[dtstationindex];}
-//protected:
-  std::map<DTStationIndex2,std::set<RPCDetId> > rollstoreDT;
-  ObjectMap2(const edm::EventSetup& iSetup);
-private:
-  static ObjectMap2* mapInstance;
-}; 
-class CSCStationIndex2{
-public:
-  CSCStationIndex2():_region(0),_station(0),_ring(0),_chamber(0){}
-  CSCStationIndex2(int region, int station, int ring, int chamber):
-    _region(region),
-    _station(station),
-    _ring(ring),
-    _chamber(chamber){}
-  ~CSCStationIndex2(){}
-  int region() const {return _region;}
-  int station() const {return _station;}
-  int ring() const {return _ring;}
-  int chamber() const {return _chamber;}
-  bool operator<(const CSCStationIndex2& cscind) const{
-    if(cscind.region()!=this->region())
-      return cscind.region()<this->region();
-    else if(cscind.station()!=this->station())
-      return cscind.station()<this->station();
-    else if(cscind.ring()!=this->ring())
-      return cscind.ring()<this->ring();
-    else if(cscind.chamber()!=this->chamber())
-      return cscind.chamber()<this->chamber();
-    return false;
-  }
-
-private:
-  int _region;
-  int _station;
-  int _ring;  
-  int _chamber;
-};
-
-class ObjectMap2CSC{
-public:
-  static ObjectMap2CSC* GetInstance(const edm::EventSetup& iSetup);
-  std::set<RPCDetId> GetRolls(CSCStationIndex2 cscstationindex){return mapInstance->rollstoreCSC[cscstationindex];}
-//protected:
-  std::map<CSCStationIndex2,std::set<RPCDetId> > rollstoreCSC;
-  ObjectMap2CSC(const edm::EventSetup& iSetup);
-private:
-  static ObjectMap2CSC* mapInstance;
-}; 
 
 #endif
