+++ RecoLocalMuon/RPCRecHit/src/CSCStationIndex.h
@@ -1,0 +1,38 @@
+#ifndef RecoLocalMuon_RPCRecHit_CSCStationIndex_h
+#define RecoLocalMuon_RPCRecHit_CSCStationIndex_h
+
+class CSCStationIndex{
+public:
+  CSCStationIndex():_region(0),_station(0),_ring(0),_chamber(0){}
+
+  CSCStationIndex(int region, int station, int ring, int chamber):
+    _region(region),
+    _station(station),
+    _ring(ring),
+    _chamber(chamber){}
+
+  int region() const {return _region;}
+  int station() const {return _station;}
+  int ring() const {return _ring;}
+  int chamber() const {return _chamber;}
+
+  bool operator<(const CSCStationIndex& cscind) const{
+    if(cscind.region()!=this->region())
+      return cscind.region()<this->region();
+    else if(cscind.station()!=this->station())
+      return cscind.station()<this->station();
+    else if(cscind.ring()!=this->ring())
+      return cscind.ring()<this->ring();
+    else if(cscind.chamber()!=this->chamber())
+      return cscind.chamber()<this->chamber();
+    return false;
+  }
+
+private:
+  int _region;
+  int _station;
+  int _ring;  
+  int _chamber;
+};
+
+#endif // RecoLocalMuon_RPCRecHit_CSCStationIndex_h
