+++ RecoLocalMuon/RPCRecHit/src/TracktoRPC.cc
@@ -9,111 +9,15 @@
 #include <Geometry/RPCGeometry/interface/RPCGeomServ.h>
 #include <DataFormats/RPCRecHit/interface/RPCRecHit.h>
 #include <DataFormats/RPCRecHit/interface/RPCRecHitCollection.h>
-//#include "RecoLocalMuon/RPCRecHit/interface/DTSegtoRPC.h"
-//#include "RecoLocalMuon/RPCRecHit/interface/CSCSegtoRPC.h"
 #include <DataFormats/DetId/interface/DetId.h>
 #include <RecoLocalMuon/RPCRecHit/interface/TracktoRPC.h>
+#include "RecoLocalMuon/RPCRecHit/src/DTStationIndex.h"
+#include "RecoLocalMuon/RPCRecHit/src/DTObjectMap.h"
+#include "RecoLocalMuon/RPCRecHit/src/CSCStationIndex.h"
+#include "RecoLocalMuon/RPCRecHit/src/CSCObjectMap.h"
+
 #include <ctime>
 #include <TMath.h>
-
-ObjectMap2* ObjectMap2::mapInstance = NULL;
-
-ObjectMap2* ObjectMap2::GetInstance(const edm::EventSetup& iSetup){
-  if (mapInstance == NULL){
-    mapInstance = new ObjectMap2(iSetup);
-  }
-  return mapInstance;
-}
-
-ObjectMap2::ObjectMap2(const edm::EventSetup& iSetup){
-  edm::ESHandle<RPCGeometry> rpcGeo;
-  edm::ESHandle<DTGeometry> dtGeo;
-  
-  iSetup.get<MuonGeometryRecord>().get(rpcGeo);
-  iSetup.get<MuonGeometryRecord>().get(dtGeo);
-  
-  for (TrackingGeometry::DetContainer::const_iterator it=rpcGeo->dets().begin();it<rpcGeo->dets().end();it++){
-    if(dynamic_cast< RPCChamber const * >( *it ) != 0 ){
-      auto ch = dynamic_cast<const RPCChamber* >( *it ); 
-      std::vector< const RPCRoll*> roles = (ch->rolls());
-      for(std::vector<const RPCRoll*>::const_iterator r = roles.begin();r != roles.end(); ++r){
-	RPCDetId rpcId = (*r)->id();
-	int region=rpcId.region();
-	if(region==0){
-	  int wheel=rpcId.ring();
-	  int sector=rpcId.sector();
-	  int station=rpcId.station();
-	  DTStationIndex2 ind(region,wheel,sector,station);
-	  std::set<RPCDetId> myrolls;
-	  if (rollstoreDT.find(ind)!=rollstoreDT.end()) myrolls=rollstoreDT[ind];
-	  myrolls.insert(rpcId);
-	  rollstoreDT[ind]=myrolls;
-	}
-      }
-    }
-  }
-}
-
-int distsector2(int sector1,int sector2){
-  if(sector1==13) sector1=4;
-  if(sector1==14) sector1=10;
-  
-  if(sector2==13) sector2=4;
-  if(sector2==14) sector2=10;
-  
-  int distance = std::abs(sector1 - sector2);
-  if(distance>6) distance = 12-distance;
-  return distance;
-}
-
-int distwheel2(int wheel1,int wheel2){
-  int distance = std::abs(wheel1 - wheel2);
-  return distance;
-}
-ObjectMap2CSC* ObjectMap2CSC::mapInstance = NULL;
-
-ObjectMap2CSC* ObjectMap2CSC::GetInstance(const edm::EventSetup& iSetup){
-  if (mapInstance == NULL){
-    mapInstance = new ObjectMap2CSC(iSetup);
-  }
-  return mapInstance;
-}
-
-ObjectMap2CSC::ObjectMap2CSC(const edm::EventSetup& iSetup){
-  edm::ESHandle<RPCGeometry> rpcGeo;
-  edm::ESHandle<CSCGeometry> cscGeo;
-  
-  iSetup.get<MuonGeometryRecord>().get(rpcGeo);
-  iSetup.get<MuonGeometryRecord>().get(cscGeo);
-  
-  for (TrackingGeometry::DetContainer::const_iterator it=rpcGeo->dets().begin();it<rpcGeo->dets().end();it++){
-    if(dynamic_cast< const RPCChamber* >( *it ) != 0 ){
-      const RPCChamber* ch = dynamic_cast< const RPCChamber* >( *it ); 
-      std::vector< const RPCRoll*> roles = (ch->rolls());
-      for(std::vector<const RPCRoll*>::const_iterator r = roles.begin();r != roles.end(); ++r){
-	RPCDetId rpcId = (*r)->id();
-	int region=rpcId.region();
-	if(region!=0){
-	  int station=rpcId.station();
-          int ring=rpcId.ring();
-          int cscring=ring;
-          int cscstation=station;
-	  RPCGeomServ rpcsrv(rpcId);
-	  int rpcsegment = rpcsrv.segment();
-	  int cscchamber = rpcsegment; //FIX THIS ACCORDING TO RPCGeomServ::segment()Definition
-          if((station==2||station==3)&&ring==3){//Adding Ring 3 of RPC to the CSC Ring 2
-            cscring = 2;
-          }
-	  CSCStationIndex2 ind(region,cscstation,cscring,cscchamber);
-          std::set<RPCDetId> myrolls;
-	  if (rollstoreCSC.find(ind)!=rollstoreCSC.end()) myrolls=rollstoreCSC[ind];
-	  myrolls.insert(rpcId);
-          rollstoreCSC[ind]=myrolls;
-	}
-      }
-    }
-  }
-}
 
 bool TracktoRPC::ValidRPCSurface(RPCDetId rpcid, LocalPoint LocalP, const edm::EventSetup& iSetup)
 {
@@ -217,8 +121,8 @@ for(trackingRecHit_iterator hit=track->recHitsBegin(); hit != track->recHitsEnd(
                 DTChamberId dtid(geomDet->geographicalId().rawId());
                 int dtW=dtid.wheel(), dtS=dtid.sector(), dtT=dtid.station();
                 if(dtS==13) dtS=4; if(dtS==14) dtS=10;
-                ObjectMap2* TheObject = ObjectMap2::GetInstance(iSetup);
-                DTStationIndex2 theindex(0,dtW,dtS,dtT);
+                DTObjectMap* TheObject = DTObjectMap::GetInstance(iSetup);
+                DTStationIndex theindex(0,dtW,dtS,dtT);
                 std::set<RPCDetId> rollsForThisDT = TheObject->GetInstance(iSetup)->GetRolls(theindex);
                 for(std::set<RPCDetId>::iterator iteraRoll = rollsForThisDT.begin();iteraRoll != rollsForThisDT.end(); iteraRoll++)
                 {                                 
@@ -264,12 +168,12 @@ for(trackingRecHit_iterator hit=track->recHitsBegin(); hit != track->recHitsEnd(
                 float dx = trajLP.x()-trackLP.x(), dy=trajLP.y()-trackLP.y();//, dz=trajLP.z()-trackLP.z();
                 if( dx>10. && dy>10.) continue;
 
-                ObjectMap2CSC* TheObjectCSC = ObjectMap2CSC::GetInstance(iSetup);
+                CSCObjectMap* TheObjectCSC = CSCObjectMap::GetInstance(iSetup);
 	        int En = cscid.endcap(), St = cscid.station(), Ri = cscid.ring();
 	        int rpcSegment = cscid.chamber();
                 if(En==2) En= -1; if(Ri==4) Ri =1; 
 
-                CSCStationIndex2 theindex(En,St,Ri,rpcSegment);
+                CSCStationIndex theindex(En,St,Ri,rpcSegment);
                 std::set<RPCDetId> rollsForThisCSC = TheObjectCSC->GetInstance(iSetup)->GetRolls(theindex);
                 for (std::set<RPCDetId>::iterator iteraRoll = rollsForThisCSC.begin();iteraRoll != rollsForThisCSC.end(); iteraRoll++)
                 {
